Created Fragment support but GestureDetector still won't work
