+++ app/src/main/res/layout/content_main.xml
@@ -1,61 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/content_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:context="com.rogowiczdawid.smartnote.MainActivity"
     tools:showIn="@layout/app_bar_main">
 
-    <EditText
+    <FrameLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/title_bar"
-        android:background="#AEC2"
-        android:text="@string/title"
-        android:textAlignment="center"
-        android:textSize="24sp"
-        android:focusable="false"
-        />
-
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_above="@+id/editText"
-        android:layout_below="@+id/title_bar">
-
-        <LinearLayout
-            android:id="@+id/mainContainer"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:paddingLeft="@dimen/activity_horizontal_margin"
-            android:paddingRight="@dimen/activity_horizontal_margin" />
-
-    </ScrollView>
-
-    <LinearLayout
-        android:id="@+id/editText"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:orientation="horizontal">
-
-        <AutoCompleteTextView
-            android:id="@+id/getTextView"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:onClick="onAdd"
-            android:text="@string/add" />
-
-    </LinearLayout>
+        android:layout_height="match_parent"
+        android:id="@+id/main_frame" />
 
 </RelativeLayout>
