+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -3,11 +3,11 @@
 
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/nav_camera"
+            android:id="@+id/nav_note"
             android:icon="@drawable/ic_menu_camera"
             android:title="Note" />
         <item
-            android:id="@+id/nav_slideshow"
+            android:id="@+id/nav_todo"
             android:icon="@drawable/ic_menu_gallery"
             android:title="To Do List" />
         <item
