+++ app/src/main/res/values/strings.xml
@@ -8,4 +8,5 @@
     <string name="delete_button">X</string>
     <string name="add">Add</string>
     <string name="title">Title</string>
+    <string name="hold_to_set_the_title">Hold to set the title</string>
 </resources>
