+++ app/src/test/java/com/rogowiczdawid/smartnote/ExampleUnitTest.java
@@ -2,7 +2,7 @@ package com.rogowiczdawid.smartnote;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Example local unit test, which will execute on the development machine (host).
